+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Update Map, World and Physics events to use safe methods for getting their inner data
 - Update Map and physics events to use `AsssetId` instead of a direct `Handle`
 - Update the `TiledMapLayer` marker component to not contains a reference to the map `AssetId`, we should instead query the map directly
+- Prevent mutating our Map and World asset for nothing: it was triggering an additionnal asset reload
+- Clear our tileset cache when we receive an `AssetEvent::Modified` (ie. the asset is reloaded from disk)
 
 ## v0.5.1